week 8 - App.js api 통신 --> nav 컴포넌트 api 통신
