+++ src/pages/FolderPage.js
@@ -1,0 +1,4 @@
+function FolderPage() {
+  return <div className="FolderPage"></div>;
+}
+export default FolderPage;
